use ReduxBase for Reader
